backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-multipart==0.0.9
aiofiles==23.2.1
python-magic==0.4.27
groq==0.4.2
pydantic==2.6.1
pydantic-settings==2.1.0
python-dotenv==1.0.1 